fix: decouple rewards component
