+++ src/modules/project/pages/projectCreate/components/ProjectCreateRewardMain.tsx
@@ -8,7 +8,7 @@ import { Body1 } from '../../../../../components/typography'
 import { getPath } from '../../../../../constants'
 import { GeyserRewardsGuideLink } from '../../../../../constants/platform/url'
 import { useProjectByNameOrIdQuery } from '../../../../../types'
-import { Rewards } from '../../projectView/views/projectMainBody/sections/Rewards'
+import { ProjectCreationRewards as Rewards } from '../components/ProjectCreationRewards'
 import useProjectRewardTemplates from '../hooks/useProjectRewardTemplates'
 import { RewardCategory } from '../types'
 import { RewardTemplate } from './RewardTemplate'
