+++ src/modules/project/pages/projectView/views/projectMainBody/sections/Rewards.tsx
@@ -1,18 +1,12 @@
-import { useMutation } from '@apollo/client'
 import { GridItem, IconButton, SimpleGrid, Text, useBreakpoint } from '@chakra-ui/react'
-import { forwardRef, useState } from 'react'
+import { forwardRef } from 'react'
 import { useTranslation } from 'react-i18next'
 import { BsPencilFill } from 'react-icons/bs'
-import { Link, useLocation, useNavigate } from 'react-router-dom'
+import { Link, useLocation } from 'react-router-dom'
 
 import { CardLayout } from '../../../../../../../components/layouts'
-import { DeleteConfirmModal } from '../../../../../../../components/molecules'
 import { TitleDivider } from '../../../../../../../components/ui/TitleDivider'
 import { getPath, ID } from '../../../../../../../constants'
-import { MUTATION_DELETE_PROJECT_REWARD } from '../../../../../../../graphql/mutations'
-import { useModal } from '../../../../../../../hooks/useModal'
-import { ProjectRewardForCreateUpdateFragment as Reward } from '../../../../../../../types'
-import { Project, ProjectRewardForCreateUpdateFragment } from '../../../../../../../types/generated/graphql'
 import { isActive, toInt, useMobileMode, useNotification } from '../../../../../../../utils'
 import { truthyFilter } from '../../../../../../../utils/array'
 import { MobileViews, useProjectContext } from '../../../../../context'
@@ -21,108 +15,23 @@ import { RewardCard } from '../components/RewardCard'
 export const Rewards = forwardRef<HTMLDivElement>((_, ref) => {
   const { t } = useTranslation()
   const isMobile = useMobileMode()
-  const navigate = useNavigate()
   const location = useLocation()
   const breakpoint = useBreakpoint({ ssr: false })
   const largeView = ['xl', '2xl'].includes(breakpoint)
   const { toast } = useNotification()
-  const isLaunch = location.pathname.includes('launch')
 
   const {
     project,
     setMobileView,
     fundForm: { state: fundFormState, setState: setFundingFormState, updateReward },
     isProjectOwner,
-    updateProject,
-    refetch,
   } = useProjectContext()
-
-  const [selectedReward, setSelectedReward] = useState<ProjectRewardForCreateUpdateFragment>()
-  const { isOpen: isRewardDeleteOpen, onClose: onRewardDeleteClose, onOpen: openRewardDelete } = useModal()
-
-  const [deleteRewardMutation] = useMutation<any, { input: { projectRewardId: Number } }>(
-    MUTATION_DELETE_PROJECT_REWARD,
-    {
-      onCompleted() {
-        const newRewards = project?.rewards?.filter((reward) => reward?.id !== selectedReward?.id)
-        updateProject({ rewards: newRewards || [] } as Project)
-        handleClose()
-        refetch()
-        toast({
-          title: 'Successfully removed!',
-          description: `${t('Reward')} ${selectedReward?.name} ${t('was successfully removed')}`,
-          status: 'success',
-        })
-      },
-      onError(error) {
-        handleClose()
-        toast({
-          title: 'Failed to remove reward',
-          description: `${error}`,
-          status: 'error',
-        })
-      },
-    },
-  )
 
   if (!project || !isActive || !project.rewards || project.rewards.length === 0) {
     return null
   }
 
   const activeProjectRewards = project.rewards.filter((reward) => reward.isHidden === false)
-
-  const handleOnRewardSelect = (reward: Reward, count: number) => {
-    const isRewardAvailable = reward.maxClaimable ? reward.maxClaimable - reward.sold > count : true
-    if (isRewardAvailable) {
-      updateReward({ id: toInt(reward.id), count: count + 1 })
-    } else {
-      toast({
-        title: 'Reward Limit',
-        description: `Maximum number of ${(reward.maxClaimable || 0) - (reward.sold || 0)} rewards are available`,
-        status: 'error',
-      })
-    }
-
-    setMobileView(MobileViews.funding)
-    setFundingFormState('step', 'contribution')
-  }
-
-  const handleClose = () => {
-    setSelectedReward(undefined)
-    onRewardDeleteClose()
-  }
-
-  const triggerRewardRemoval = (id?: number) => {
-    const currentReward = project.rewards?.find((reward) => reward?.id === id)
-    if (!currentReward) {
-      return
-    }
-
-    setSelectedReward(currentReward)
-    openRewardDelete()
-  }
-
-  const handleRemoveReward = async () => {
-    if (!selectedReward?.id) {
-      return
-    }
-
-    deleteRewardMutation({
-      variables: {
-        input: {
-          projectRewardId: selectedReward.id,
-        },
-      },
-    })
-  }
-
-  const handleOnRewardEdit = (reward: Reward) => {
-    navigate(getPath('launchProjectRewardsEdit', project.id, reward.id))
-  }
-
-  const handleOnRewardDelete = (reward: Reward) => {
-    triggerRewardRemoval(reward.id)
-  }
 
   const renderRewards = () => {
     if (activeProjectRewards.length > 0) {
@@ -134,9 +43,21 @@ export const Rewards = forwardRef<HTMLDivElement>((_, ref) => {
             width="100%"
             reward={reward}
             count={count}
-            onRewardClick={() => handleOnRewardSelect(reward, count)}
-            handleEdit={() => handleOnRewardEdit(reward)}
-            handleRemove={() => handleOnRewardDelete(reward)}
+            onRewardClick={() => {
+              const isRewardAvailable = reward.maxClaimable ? reward.maxClaimable - reward.sold > count : true
+              if (isRewardAvailable) {
+                updateReward({ id: toInt(reward.id), count: count + 1 })
+              } else {
+                toast({
+                  title: 'Reward Limit',
+                  description: `Maximum number of ${reward.maxClaimable - reward.sold} rewards are available`,
+                  status: 'error',
+                })
+              }
+
+              setMobileView(MobileViews.funding)
+              setFundingFormState('step', 'contribution')
+            }}
           />
         )
       })
@@ -175,29 +96,15 @@ export const Rewards = forwardRef<HTMLDivElement>((_, ref) => {
         alignItems="flex-start"
         spacing="25px"
         mobileDense
-        {...(isLaunch ? { border: 'none', p: 0 } : {})}
       >
-        {!isLaunch && (
-          <TitleDivider
-            badge={activeProjectRewards.length}
-            isFixed={isRewardTitleFixed}
-            rightAction={rightActionButton}
-          >
-            {t('Rewards')}
-          </TitleDivider>
-        )}
+        <TitleDivider badge={activeProjectRewards.length} isFixed={isRewardTitleFixed} rightAction={rightActionButton}>
+          {t('Rewards')}
+        </TitleDivider>
 
         <SimpleGrid columns={largeView ? 2 : 1} spacing={'20px'} width={'100%'}>
           {renderRewards()}
         </SimpleGrid>
       </CardLayout>
-      <DeleteConfirmModal
-        isOpen={isRewardDeleteOpen}
-        onClose={handleClose}
-        title={`${t('Delete reward')}`}
-        description={t('Are you sure you want to remove the reward?')}
-        confirm={handleRemoveReward}
-      />
     </>
   )
 })
